pyproject.toml
[tool.poetry]
name = "sentimentops"
version = "0.1.0"
description = "End-to-end NLP pipeline with medallion architecture, MLOps, and REST API serving for sentiment classification on Amazon review data"
authors = ["AqeelMadhi <aqeelmadhi25@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
# Root dependencies — needed across every layer
python = "^3.10"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"

[tool.poetry.group.data.dependencies]
# Data engineering layer — pipelines, transforms, validation
pandas = "^2.0.0"
polars = "^0.20.0"
duckdb = "^0.10.0"
deltalake = "^0.17.0"
great-expectations = "^0.18.0"
dvc = "^3.0.0"
feast = "^0.38.0"
pyarrow = "^15.0.0"
requests = "^2.31.0"
datasets = "^2.18.0"
dvc-azure = "^3.1.0"


[tool.poetry.group.ml.dependencies]
# ML/training layer — modelling, tuning, explainability
numpy = "^1.26.0"
scikit-learn = "^1.4.0"
transformers = "^4.38.0"
torch = "^2.2.0"
optuna = "^3.5.0"
shap = "^0.44.0"
mlflow = "^2.11.0"
evaluate = "^0.4.0"
accelerate = "^0.27.0"

[tool.poetry.group.api.dependencies]
# Serving layer — FastAPI, caching, monitoring
fastapi = "^0.110.0"
uvicorn = {version = "^0.27.0", extras = ["standard"]}
pydantic = "^2.6.0"
redis = "^5.0.0"
prometheus-fastapi-instrumentator = "^6.1.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
# Dev tooling — testing, linting, formatting
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"
ruff = "^0.3.0"
pre-commit = "^3.6.0"
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["N", "B", "UP", "E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]
